+++ src/main/java/com/amazonaws/services/neptune/propertygraph/io/QueryTask.java
@@ -14,14 +14,12 @@ package com.amazonaws.services.neptune.propertygraph.io;
 
 import com.amazonaws.services.neptune.io.Directories;
 import com.amazonaws.services.neptune.io.Status;
-import com.amazonaws.services.neptune.propertygraph.AllLabels;
-import com.amazonaws.services.neptune.propertygraph.EdgeLabelStrategy;
+import com.amazonaws.services.neptune.propertygraph.ExportStats;
 import com.amazonaws.services.neptune.propertygraph.Label;
 import com.amazonaws.services.neptune.propertygraph.LabelsFilter;
 import com.amazonaws.services.neptune.propertygraph.NamedQuery;
 import com.amazonaws.services.neptune.propertygraph.NeptuneGremlinClient;
-import com.amazonaws.services.neptune.propertygraph.NodeLabelStrategy;
-import com.amazonaws.services.neptune.propertygraph.io.result.PGEdgeResult;
+import com.amazonaws.services.neptune.propertygraph.StatsContainer;
 import com.amazonaws.services.neptune.propertygraph.io.result.QueriesEdgeResult;
 import com.amazonaws.services.neptune.propertygraph.io.result.QueriesNodeResult;
 import com.amazonaws.services.neptune.propertygraph.schema.FileSpecificLabelSchemas;
@@ -57,6 +55,7 @@ public class QueryTask implements Callable<Map<GraphElementType, FileSpecificLab
     private final boolean structuredOutput;
     private final LabelsFilter nodeLabelFilter;
     private final LabelsFilter edgeLabelFilter;
+    private final ExportStats exportStats;
 
     public QueryTask(Queue<NamedQuery> queries,
                      NeptuneGremlinClient.QueryClient queryClient,
@@ -67,7 +66,8 @@ public class QueryTask implements Callable<Map<GraphElementType, FileSpecificLab
                      AtomicInteger index,
                      boolean structuredOutput,
                      LabelsFilter nodeLabelFilter,
-                     LabelsFilter edgeLabelFilter) {
+                     LabelsFilter edgeLabelFilter,
+                     ExportStats exportStats) {
 
         this.queries = queries;
         this.queryClient = queryClient;
@@ -79,6 +79,7 @@ public class QueryTask implements Callable<Map<GraphElementType, FileSpecificLab
         this.structuredOutput = structuredOutput;
         this.nodeLabelFilter = nodeLabelFilter;
         this.edgeLabelFilter = edgeLabelFilter;
+        this.exportStats = exportStats;
     }
 
     @Override
@@ -163,8 +164,8 @@ public class QueryTask implements Callable<Map<GraphElementType, FileSpecificLab
                                 graphElementSchemas,
                                 targetConfig,
                                 (WriterFactory<QueriesNodeResult>) GraphElementType.nodes.writerFactory(),
-                                new LabelWriters<>(new AtomicInteger(),0),
-                                null,
+                                new LabelWriters<>(new AtomicInteger(), 0),
+                                new ExportStatsWrapper(exportStats, GraphElementType.nodes),
                                 status,
                                 index,
                                 nodeLabelFilter)
@@ -175,8 +176,8 @@ public class QueryTask implements Callable<Map<GraphElementType, FileSpecificLab
                                     graphElementSchemas,
                                     targetConfig,
                                     (WriterFactory<QueriesEdgeResult>) GraphElementType.edges.writerFactory(),
-                                    new LabelWriters<>(new AtomicInteger(),0),
-                                    null,
+                                    new LabelWriters<>(new AtomicInteger(), 0),
+                                    new ExportStatsWrapper(exportStats, GraphElementType.edges),
                                     status,
                                     index,
                                     edgeLabelFilter)
@@ -321,6 +322,25 @@ public class QueryTask implements Callable<Map<GraphElementType, FileSpecificLab
 
         private QueriesEdgeResult getQueriesEdgeResult(Map<?, ?> map) {
             return new QueriesEdgeResult(map);
+        }
+    }
+
+    private class ExportStatsWrapper implements StatsContainer {
+        private final ExportStats exportStats;
+        private final GraphElementType graphElementType;
+
+        public ExportStatsWrapper(ExportStats exportStats, GraphElementType graphElementType) {
+            this.exportStats = exportStats;
+            this.graphElementType = graphElementType;
+        }
+
+        @Override
+        public void updateStats(Label label) {
+            if(graphElementType.equals(GraphElementType.nodes)) {
+                exportStats.incrementNodeStats(label);
+            } else {
+                exportStats.incrementEdgeStats(label);
+            }
         }
     }
 }