+++ src/test/java/com/amazonaws/services/neptune/ExportPgFromQueriesIntegrationTest.java
@@ -1,14 +1,9 @@
 package com.amazonaws.services.neptune;
 
 import com.amazonaws.services.neptune.export.NeptuneExportRunner;
-import com.amazonaws.services.neptune.propertygraph.io.JsonResource;
-import com.amazonaws.services.neptune.propertygraph.schema.GraphSchema;
 import org.junit.Test;
 
 import java.io.File;
-import java.io.IOException;
-import java.net.URI;
-import java.net.URISyntaxException;
 
 import static org.junit.Assert.assertTrue;
 
@@ -48,6 +43,21 @@ public class ExportPgFromQueriesIntegrationTest extends AbstractExportIntegratio
     public void testExportPgFromQueriesWithStructuredOutputWithEdgeAndVertexLabels() {
         final String[] command = {"export-pg-from-queries", "-e", neptuneEndpoint,
                 "-d", outputDir.getPath(),
+                "-q", "all=g.V().union(elementMap(), outE().elementMap())",
+                "--edge-label-strategy", "edgeAndVertexLabels", "--structured-output"
+        };
+        final NeptuneExportRunner runner = new NeptuneExportRunner(command);
+        runner.run();
+
+        final File resultDir = outputDir.listFiles()[0];
+
+        assertEquivalentStructuredOutput(new File("src/test/resources/IntegrationTest/testExportPgWithEdgeAndVertexLabelsWithoutTypes"), resultDir);
+    }
+
+    @Test
+    public void testExportPgFromQueriesWithStructuredOutputWithEdgeAndVertexLabelsIncludeTypes() {
+        final String[] command = {"export-pg-from-queries", "-e", neptuneEndpoint,
+                "-d", outputDir.getPath(),
                 "-q", "airport=g.V().union(hasLabel('airport'), outE()).elementMap()",
                 "--include-type-definitions", "--edge-label-strategy", "edgeAndVertexLabels",
                 "--structured-output"
@@ -57,12 +67,15 @@ public class ExportPgFromQueriesIntegrationTest extends AbstractExportIntegratio
 
         final File resultDir = outputDir.listFiles()[0];
 
-        assertEquivalentStructuredOutput(new File("src/test/resources/IntegrationTest/testExportPgFromQueriesStructuredOutput"), resultDir);
+        assertEquivalentStructuredOutput(new File("src/test/resources/IntegrationTest/testExportPgFromQueriesStructuredOutputWithEdgeAndVertexLabels"), resultDir);
     }
 
     @Override
     protected void assertEquivalentResults(final File expected, final File actual) {
-        assertTrue("queries.json does not match expected results", areJsonContentsEqual(expected.listFiles((dir, name) -> name.equals("queries.json"))[0], actual.listFiles((dir, name) -> name.equals("queries.json"))[0]));
+        assertJSONContentMatches(
+                expected.listFiles((dir, name) -> name.equals("queries.json"))[0],
+                actual.listFiles((dir, name) -> name.equals("queries.json"))[0],
+                "queries.json does not match expected results");
         for (File expectedResultsDir : expected.listFiles((dir, name) -> name.equals("results"))[0].listFiles()) {
             assertTrue(expectedResultsDir.isDirectory());
             String dirName = expectedResultsDir.getName();
@@ -71,21 +84,7 @@ public class ExportPgFromQueriesIntegrationTest extends AbstractExportIntegratio
     }
 
     private void assertEquivalentStructuredOutput(final File expected, final File actual) {
-        GraphSchema config = null;
-        try {
-            config = new JsonResource<GraphSchema, Boolean>(
-                    "Config file",
-                    new URI(expected.getPath() + "/config.json"),
-                    GraphSchema.class).get();
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        } catch (URISyntaxException e) {
-            throw new RuntimeException(e);
-        }
-
-        if (expected.listFiles(((dir, name) -> name.equals("nodes"))).length >= 1) {
-            assertTrue("nodes directory does not match expected results", areDirContentsEquivalent(expected + "/nodes", actual + "/nodes", config));
-        }
+        super.assertEquivalentResults(expected, actual);
     }