+++ src/test/java/com/amazonaws/services/neptune/propertygraph/io/QueryTaskTest.java
@@ -1,15 +1,35 @@
+/*
+Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+Licensed under the Apache License, Version 2.0 (the "License").
+You may not use this file except in compliance with the License.
+A copy of the License is located at
+    http://www.apache.org/licenses/LICENSE-2.0
+or in the "license" file accompanying this file. This file is distributed
+on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+express or implied. See the License for the specific language governing
+permissions and limitations under the License.
+*/
+
 package com.amazonaws.services.neptune.propertygraph.io;
 
+import com.amazonaws.services.neptune.export.FeatureToggles;
 import com.amazonaws.services.neptune.io.Status;
 import com.amazonaws.services.neptune.io.StatusOutputFormat;
 import com.amazonaws.services.neptune.propertygraph.AllLabels;
 import com.amazonaws.services.neptune.propertygraph.EdgeLabelStrategy;
+import com.amazonaws.services.neptune.propertygraph.ExportStats;
+import com.amazonaws.services.neptune.propertygraph.GremlinFilters;
 import com.amazonaws.services.neptune.propertygraph.Label;
 import com.amazonaws.services.neptune.propertygraph.NamedQuery;
 import com.amazonaws.services.neptune.propertygraph.NeptuneGremlinClient;
 import com.amazonaws.services.neptune.propertygraph.NodeLabelStrategy;
+import com.amazonaws.services.neptune.propertygraph.schema.ExportSpecification;
 import com.amazonaws.services.neptune.propertygraph.schema.FileSpecificLabelSchemas;
+import com.amazonaws.services.neptune.propertygraph.schema.GraphElementSchemas;
 import com.amazonaws.services.neptune.propertygraph.schema.GraphElementType;
+import com.amazonaws.services.neptune.propertygraph.schema.GraphSchema;
+import com.amazonaws.services.neptune.propertygraph.schema.MasterLabelSchemas;
+import com.fasterxml.jackson.databind.JsonNode;
 import org.apache.tinkerpop.gremlin.driver.Result;
 import org.apache.tinkerpop.gremlin.driver.ResultSet;
 import org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.GraphTraversal;
@@ -19,6 +39,10 @@ import org.apache.tinkerpop.gremlin.tinkergraph.structure.TinkerFactory;
 import org.junit.Test;
 
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.Map;
 import java.util.Queue;
@@ -79,6 +103,50 @@ public class QueryTaskTest {
         assertTrue(edgeSchemas.hasSchemasForLabel(new Label("created")));
     }
 
+    @Test
+    public void shouldUpdateExportStats() throws Exception {
+        ExportStats exportStats = new ExportStats();
+        QueryTask qt = createQueryTask(gModern.V().union(__.elementMap(), __.outE().elementMap()), true, exportStats);
+
+        Map<GraphElementType, FileSpecificLabelSchemas> results = qt.call();
+
+        Map<GraphElementType, GraphElementSchemas> graphElementSchemas = new HashMap<>();
+        Collection<ExportSpecification> exportSpecifications = createExportSpecifications(exportStats);
+
+        for(ExportSpecification exportSpecification : exportSpecifications) {
+            MasterLabelSchemas masterLabelSchemas = exportSpecification.createMasterLabelSchemas(
+                    Collections.singletonList(results.get(exportSpecification.getGraphElementType()))
+            );
+            try {
+                graphElementSchemas.put(exportSpecification.getGraphElementType(), masterLabelSchemas.toGraphElementSchemas());
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+        }
+
+        JsonNode statsResults = exportStats.toJson(new GraphSchema(graphElementSchemas)).findPath("stats");
+        assertEquals(6, statsResults.findValue("nodes").intValue());
+        assertEquals(6, statsResults.findValue("edges").intValue());
+        //NOTE: Property stats are normally recorded through the PropertyGraphPrinter, which is mocked out in this test.
+        assertEquals(0, statsResults.findValue("properties").intValue());
+
+        JsonNode nodeDetails = statsResults.findPath("details").findPath("nodes");
+        assertTrue(nodeDetails.isArray());
+        assertEquals(2, nodeDetails.size());
+        assertEquals("software", nodeDetails.get(0).get("description").textValue());
+        assertEquals(2, nodeDetails.get(0).get("count").intValue());
+        assertEquals("person", nodeDetails.get(1).get("description").textValue());
+        assertEquals(4, nodeDetails.get(1).get("count").intValue());
+
+        JsonNode edgeDetails = statsResults.findPath("details").findPath("edges");
+        assertTrue(edgeDetails.isArray());
+        assertEquals(2, edgeDetails.size());
+        assertEquals("knows", edgeDetails.get(0).get("description").textValue());
+        assertEquals(2, edgeDetails.get(0).get("count").intValue());
+        assertEquals("created", edgeDetails.get(1).get("description").textValue());
+        assertEquals(4, edgeDetails.get(1).get("count").intValue());
+    }
+
     private NeptuneGremlinClient.QueryClient getMockClient(GraphTraversal traversal) {
         NeptuneGremlinClient.QueryClient mockClient = mock(NeptuneGremlinClient.QueryClient.class);
         ResultSet results = mock(ResultSet.class);
@@ -90,6 +158,10 @@ public class QueryTaskTest {
     }
 
     private QueryTask createQueryTask(GraphTraversal traversal, boolean structuredOutput) throws IOException {
+        return createQueryTask(traversal, structuredOutput, new ExportStats());
+    }
+
+    private QueryTask createQueryTask(GraphTraversal traversal, boolean structuredOutput, ExportStats exportStats) throws IOException {
         Queue<NamedQuery> mockQueries = new LinkedList<>();
         mockQueries.add(mock(NamedQuery.class));
 
@@ -106,8 +178,30 @@ public class QueryTaskTest {
                 new AtomicInteger(),
                 structuredOutput,
                 new AllLabels(NodeLabelStrategy.nodeLabelsOnly),
-                new AllLabels(EdgeLabelStrategy.edgeLabelsOnly)
+                new AllLabels(EdgeLabelStrategy.edgeLabelsOnly),
+                exportStats
         );
+    }
+
+    private Collection<ExportSpecification> createExportSpecifications(ExportStats exportStats) {
+        Collection<ExportSpecification> exportSpecifications = new ArrayList<>();
+        exportSpecifications.add(new ExportSpecification(
+                GraphElementType.nodes,
+                new AllLabels(NodeLabelStrategy.nodeLabelsOnly),
+                GremlinFilters.EMPTY,
+                exportStats,
+                false,
+                new FeatureToggles(Collections.EMPTY_SET)
+        ));
+        exportSpecifications.add(new ExportSpecification(
+                GraphElementType.edges,
+                new AllLabels(NodeLabelStrategy.nodeLabelsOnly),
+                GremlinFilters.EMPTY,
+                exportStats,
+                false,
+                new FeatureToggles(Collections.EMPTY_SET)
+        ));
+        return exportSpecifications;
     }
 
 }